requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
constantly==15.1.0
crispy==0.7.3
cryptography==35.0.0
cssselect==1.1.0
cycler==0.10.0
Django==3.2.6
django-crispy-forms==1.12.0
django-rest-framework==0.1.0
django-rq==2.4.1
djangorestframework==3.12.4
fabio==0.12.0
Factory==1.2
factory-boy==3.2.0
Faker==9.3.1
h2==3.2.0
h5py==3.4.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
packaging==21.0
parsel==1.6.0
Pillow==8.3.2
pluggy==1.0.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
queuelib==1.6.2
redis==3.5.3
requests==2.26.0
rq==1.10.0
Scrapy==2.5.1
service-identity==21.1.0
silx==0.15.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.6
w3lib==1.22.0
zope.interface==5.4.0
